feat(alliance): 新增专家注册中心服务骨架
- 新增 mox-alliance-registry-svc crate 并注册到 workspace
- 补充 docker-compose.alliance.yml 部署编排与端口登记
- 修复 Cargo.toml 中无效的 build.target-dir 配置，迁移至 .cargo/config.toml
- 更新架构文档与基准测试结果

+++ platform/domains/kg/svc/mox-kg-service-svc/src/graph_server.rs
@@ -223,7 +223,7 @@ fn ngql_result_for(node: &PlanNode) -> ResultSet {
         GoSteps(n) => (
             vec!["steps".into(), "dst".into()],
             vec![
-                PropValue::Int(*n as i64),
+                PropValue::Int(*n),
                 PropValue::Str(format!("sink_{n}")),
             ],
             "GO STEP",
